Add support for aliased and nested factories.
