+++ CHANGELOG.markdown
@@ -1,7 +1,8 @@
 # Changelog
 
-## 0.2.0 — Unreleased.
+## 0.2.0 — Unreleased
 
+* Add support for aliased and nested factories.
 * Support multiple versions of factory_bot.
 
 ## 0.1.0 — November 5th, 2021
